base model - initial impl, moving code to a common area, extended by LogReg for now (and add some params to logreg param grid)
